requirements.txt
requests>=2.28.2
schedule>=1.1.0
python-dotenv>=0.21.0
flask>=2.2.3
flask-cors>=3.0.10
gunicorn>=20.1.0